requirements/development.txt
-r base.txt

pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=21.0.0
django-debug-toolbar>=4.2.0
ipython>=8.18.1
black>=23.12.1
isort>=5.13.2
flake8>=6.1.0
mypy>=1.7.1
django-stubs>=4.2.7
pre-commit>=3.6.0
